test(implement-loop-commit-hardening): satisfy arch gates on PR #2662
Fix three arch-adversarial (arch_shard_3) failures the mission's new test
files introduced, surfaced only on PR CI:

- golden-count ban: annotate the 7 new `len(...) == N` sites as
  `# golden-count: cardinality-is-contract` — each is a genuine cardinality
  precondition whose *which* (partition kind / tuple element / tail bound) is
  asserted on the following lines, so a content-equality would carry no more
  information. Keeps `tests/specify_cli` under its frozen ceiling without a
  baseline bump.
- marker correctness (Rule 1 git_repo presence + Rule 2 no-fast-on-subprocess):
  extract the one real-git subprocess class
  (`TestFlatLegacyNoneAtSeamReachesSuccessArms`) out of the fast/unit
  `test_partition_authority_characterization.py` into a sibling
  `test_flat_legacy_none_seam_success_arms.py` carrying
  `[integration, git_repo]`; the characterization file stays a pure `[unit,
  fast]` gate.

No behavior change; 331 touched tests + all three arch gates green locally.
