+++ tests/specify_cli/coordination/test_partition_authority_characterization.py
@@ -170,7 +170,7 @@ class TestDisagreementSetRoutesPrimaryOnResidueButCoordUnderCommitRouter:
         groups = commit_router._group_files_by_partition(
             Path("/tmp"), (Path(kind_none_path),), "m", kind=_COORD_CALLER_KIND
         )
-        assert len(groups) == 1
+        assert len(groups) == 1  # golden-count: cardinality-is-contract (single-partition; kind asserted below)
         group_kind, group_files = groups[0]
         assert is_primary_artifact_kind(group_kind), (
             f"{kind_none_path!r} (kind=None) must route to a PRIMARY-partition "
@@ -180,116 +180,6 @@ class TestDisagreementSetRoutesPrimaryOnResidueButCoordUnderCommitRouter:
         )
         assert group_kind != _COORD_CALLER_KIND
         assert Path(kind_none_path) in group_files
-
-
-class TestFlatLegacyNoneAtSeamReachesSuccessArms:
-    """T017 -- confirm WP01's (#2648) fail-close scope is the NARROW triple
-    only. A real flat/legacy mission's ``placement_ref=None`` -- with NO
-    ``coordination_branch`` in meta.json at all -- must still reach the
-    ``755`` SUCCESS arm, not the narrow-triple fail-close. This is the
-    #2463 None-overload guard (INV-7): ``placement_ref is None`` is NOT
-    unconditionally degenerate.
-    """
-
-    @staticmethod
-    def _git(repo: Path, *args: str) -> str:
-        import subprocess
-
-        return subprocess.run(
-            ["git", *args], cwd=repo, check=True, capture_output=True, text=True
-        ).stdout.strip()
-
-    def _init_repo(self, repo: Path, *, branch: str) -> None:
-        repo.mkdir()
-        self._git(repo, "init", "-q", "-b", branch)
-        self._git(repo, "config", "user.email", "t@example.com")
-        self._git(repo, "config", "user.name", "Test")
-        self._git(repo, "config", "commit.gpgsign", "false")
-        (repo / "seed.txt").write_text("seed\n", encoding="utf-8")
-        self._git(repo, "add", "seed.txt")
-        self._git(repo, "commit", "-q", "-m", "initial")
-
-    def test_flat_legacy_placement_ref_none_commits_successfully_755(
-        self, tmp_path: Path, monkeypatch: pytest.MonkeyPatch
-    ) -> None:
-        import json
-
-        from specify_cli.cli.commands.implement import (
-            _commit_planning_artifacts_transaction,
-        )
-
-        planning_branch = "mission/2650-wp04-flat-legacy-demo"
-        repo = tmp_path / "repo"
-        self._init_repo(repo, branch=planning_branch)
-
-        mission_slug = "wp04-flat-legacy-demo"
-        feature_dir = repo / "kitty-specs" / mission_slug
-        feature_dir.mkdir(parents=True)
-        # No "coordination_branch" key at all -- a genuinely flat/legacy
-        # mission (not merely a coord mission with an unprotected branch).
-        (feature_dir / "meta.json").write_text(
-            json.dumps(
-                {
-                    "mission_id": "01J9WP04FLATLEGACYXXXXXXXX",
-                    "mission_slug": mission_slug,
-                    "mid8": "01J9WP04",
-                    "mission_type": "software-dev",
-                    "target_branch": planning_branch,
-                    "created_at": "2026-07-15T00:00:00+00:00",
-                    "friendly_name": "WP04 flat/legacy None-at-seam",
-                },
-                indent=2,
-            ),
-            encoding="utf-8",
-        )
-        (feature_dir / "spec.md").write_text("# Spec\n", encoding="utf-8")
-
-        calls: list[tuple[str, list[str]]] = []
-
-        class _FakeTxn:
-            def __init__(self, destination_ref: str) -> None:
-                self._destination_ref = destination_ref
-                self._paths: list[str] = []
-
-            def __enter__(self) -> _FakeTxn:
-                return self
-
-            def __exit__(self, *_exc: object) -> None:
-                calls.append((self._destination_ref, list(self._paths)))
-
-            def write_artifact(self, repo_path: Path, _content: bytes) -> None:
-                self._paths.append(repo_path.as_posix())
-
-            def commit(self, _msg: str) -> None:
-                return None
-
-        class _FakeBookkeepingTransaction:
-            @classmethod
-            def acquire(cls, **kwargs: object) -> _FakeTxn:
-                return _FakeTxn(str(kwargs["destination_ref"]))
-
-        monkeypatch.setattr(
-            "specify_cli.coordination.transaction.BookkeepingTransaction",
-            _FakeBookkeepingTransaction,
-        )
-
-        spec_rel = f"kitty-specs/{mission_slug}/spec.md"
-
-        # NOT the narrow triple: no coord_branch at all -- must not raise.
-        _commit_planning_artifacts_transaction(
-            repo_root=repo,
-            feature_dir=feature_dir,
-            mission_slug=mission_slug,
-            planning_branch=planning_branch,
-            files_to_commit=[spec_rel],
-            commit_msg="chore: flat/legacy None-at-seam characterization",
-            placement_ref=None,
-        )
-
-        assert calls == [(planning_branch, [spec_rel])], (
-            "a flat/legacy mission's placement_ref=None must reach the 755 "
-            "SUCCESS arm (C-004 strangler), not the narrow-triple fail-close"
-        )
 
 
 class TestIntendedUnifiedContractCliSideOnly:
